Add the mixed simulation reasoning principle
The mixed simulation reasoning principle lets you (locally) do forward
reasoning even if the target language as a whole is not determinate.
It is required only for the target *state*, not the entire language, to
be determinate.  This principle logically subsumes both the forawrd
and backward reasoning principle.

This principle is theoretically motivated by:

  Neis et al.  Pilsner: A Compositionally Verified Compiler for a Higher-Order Imperative Language.  ICFP 2015.

The main motivation for this principle is to fully support integer-
pointer casts in CompCert.  The following paper introduces such a model,
yet the model is nondeterministic and is not fit for the existing proof
structure of CompCert based on the forward simulation reasoning
principle.   The mixed simulation reasoning principle solves the problem
by supporting both the backward reasoning and the forward reasoning (for
determinate states).

  Kang et al.  A Formal C Memory Model Supporting Integer-Pointer Casts.  PLDI 2015.

The second motivastion is to remove axioms on external calls.  CompCert
assumes that external calls are (1) "receptive" and (2) "determinate"
(variants of determinism; defined in common/Smallstep.v).  These axioms
are introduced to enable us to use forward reasoning.  We want to lift
this restriction by using mixed simulation, or in other words, backward
reasoning for external calls and forward reasoning for the other steps.